Commit modification of all model + Show likes/comments/pp
I had to modify all the models in the service side. In first time, we used JsonbTransient tag to avoid infinite looping when serializing the json for the request. But we lost too many information. I had to change filed in order to define foreign key by id insted of objects. Whit this solution we avoid infinite looping error but we still get information.